% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{addPrefix}
\alias{addPrefix}
\title{Add prefix to files in a folder}
\usage{
addPrefix(prefix, path = "instances", recursiv = TRUE)
}
\arguments{
\item{prefix}{Prefix to add.}

\item{path}{Subfolder path.}

\item{recursiv}{Change files in subfolders too.}
}
\description{
Add prefix to files in a folder
}
