% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkContribution}
\alias{checkContribution}
\title{Check if a contribution convey with the guidelines.}
\usage{
checkContribution(validateResults = TRUE)
}
\arguments{
\item{validateResults}{If true the result files are validated against schema.}
}
\description{
You must run this function in the root of your contribution.
}
\author{
Lars Relund \email{lars@relund.dk}
}
