% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkMeta}
\alias{checkMeta}
\title{Validate meta file based on schema.}
\usage{
checkMeta(meta)
}
\arguments{
\item{meta}{Name of meta file.}
}
\value{
Warnings and errors (if any).
}
\description{
Validate meta file based on schema.
}
\author{
Lars Relund \email{lars@relund.dk}
}
