% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{checkResult}
\alias{checkResult}
\title{Validate a result file based on schema.}
\usage{
checkResult(file)
}
\arguments{
\item{file}{Name of result file (with path).}
}
\value{
Warnings and errors (if any).
}
\description{
Validate a result file based on schema.
}
\examples{
\dontrun{
checkResult("Tuyttens00_AP_n05_result.json")
}
}
\author{
Lars Relund \email{lars@relund.dk}
}
