% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createResultFile}
\alias{createResultFile}
\title{Create a result file}
\usage{
createResultFile(instanceName, contributionName, objectives, points, card,
  suppCard = NULL, extCard = NULL, objectiveType = rep("int", objectives),
  direction = rep("min", objectives), comments = NULL, optimal = TRUE,
  cpu = NULL, valid = TRUE, version = "1.0", other = "", misc = NULL)
}
\arguments{
\item{instanceName}{Name of instance not including the file extension (string).}

\item{contributionName}{The name of the the contribution (string).}

\item{objectives}{Number of objectives (number).}

\item{points}{A data frame with the nondominated points. Column names must be \code{z1, z2, ..., zn,
type} where column \code{zi} is the value of the i'th objective and column type contains
strings which may be either \code{us} (unsupported), \code{se} (supported extreme), \code{s}
(supported - may be extreme or nonextreme), \code{sne} (supported nonextreme), \code{NA}
(unknown).}

\item{card}{Number of points.}

\item{suppCard}{Number of supported nondominated points (number). This is with respect to the
solution found.}

\item{extCard}{Number of extreme supported nondominated points.}

\item{objectiveType}{Vector with strings \code{int}, \code{float} or \code{NA} (unknown) giving
the numeric types of the objectives. Must have the same length as the number of objectives.}

\item{direction}{Vector with strings \code{min} or \code{max} giving the direction of the
objectives.}

\item{comments}{Misc comments about the results (string).}

\item{optimal}{\code{TRUE} if an exact optimal solution, \code{FALSE} is know an approximation,
\code{NA} if unknown, i.e. may be optimal (boolean, null).}

\item{cpu}{If not NULL, a vector with two entries cpu time in seconds (number) and machine
specification (string).}

\item{valid}{If TRUE the results are considered valid. If FALSE the results may be in conflict
with results on the same instance from other contributions.}

\item{version}{Result format version (string). Currently version must be set to 1.0.}

\item{other}{String added to the result file name.}

\item{misc}{If not NULL, an entry you may use as you like (vector, list etc.). It could e.g.
contain an object with more detailed entries about the experiment.}
}
\description{
Create a result file named \code{<instanceName>_<other>_result.json} in json format.
}
\examples{
points <- data.frame(z1 = c(27, 30, 31, 34, 42, 43, 49, 51), z2 = c(56, 53, 36, 33, 30, 25, 23, 9),
   type = c('se', 'us', 'se', 'us', 'us', 'us', 'us', 'se'))
createResultFile(instanceName = "Tuyttens00_AP_n05", contributionName = "Pedersen08",
   objectives = 2, points = points, card = 8, suppCard = 3, extCard = 3,
   objectiveType = c("int", "int"), direction = c("min", "min"),
   comments = "Results from the paper by Pedersen et. al (2008)", optimal = TRUE
)
}
\author{
Lars Relund \email{lars@relund.dk}
}
