% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getContributionAsZip}
\alias{getContributionAsZip}
\title{Download a contribution as a zip file.}
\usage{
getContributionAsZip(contributionName)
}
\arguments{
\item{contributionName}{Contribution name(s) (without the MOrepo prefix).}
}
\description{
Download a contribution as a zip file.
}
\examples{
getContributionAsZip(c("Tuyttens00", "Template"))
}
\author{
Lars Relund \email{lars@relund.dk}
}
