% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getContributorNames}
\alias{getContributorNames}
\title{Get info about contributors (authors in citation BibTeX files)}
\usage{
getContributorNames(local = FALSE)
}
\arguments{
\item{local}{Use local repo.}
}
\value{
A vector with unique contributors (read from citation.bib).
}
\description{
Get info about contributors (authors in citation BibTeX files)
}
\examples{
getContributorNames()
}
\author{
Lars Relund \email{lars@relund.dk}
}
