% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getContributors}
\alias{getContributors}
\title{Get info about contributors (authors in citation BibTeX files)}
\usage{
getContributors(branch = "master", local = FALSE)
}
\arguments{
\item{branch}{The branch at GitHub.}

\item{local}{Use local repo.}
}
\value{
A vector with unique contributors.
}
\description{
Get info about contributors (authors in citation BibTeX files)
}
\examples{
getContributors()
getContributors(local = F)
}
\author{
Lars Relund \email{lars@relund.dk}
}
