% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getFileList}
\alias{getFileList}
\title{Get list of files at GitHub}
\usage{
getFileList(name = "", subdir = "", branch = "master", local = FALSE)
}
\arguments{
\item{name}{Name of the file(s) or only parts of the name. May be an regular expression.}

\item{subdir}{Restricts search to a specific subfolder path at GitHub. The path must end with a
slash (/).}

\item{branch}{The branch at GitHub.}

\item{local}{Use local repo.}
}
\value{
The names of the files (including file path at GitHub)
}
\description{
Get list of files at GitHub
}
\examples{
getFileList("SSCFLP.*p6")
getFileList("inst")
getFileList(subdir = "instances/Gadegaard16/")
}
\author{
Lars Relund \email{lars@relund.dk}
}
