% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getFilePath}
\alias{getFilePath}
\title{File path of a specific file at MOrepo.}
\usage{
getFilePath(file, contribution, local = FALSE)
}
\arguments{
\item{file}{Relative file path within the contribution.}

\item{contribution}{Contribution name.}

\item{local}{Use local repo (urls becomes paths).}
}
\value{
A vector of strings.
}
\description{
File path of a specific file at MOrepo.
}
\examples{
getFilePath("instances/raw/SSCFLP/Gadegaard16_SSCFLP_Holmberg_p64_0.raw",
             contribution = "Gadegaard16")
}
\author{
Lars Relund \email{lars@relund.dk}
}
