% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getInstance}
\alias{getInstance}
\title{Download instances from GitHub}
\usage{
getInstance(name = NULL, class = NULL, fileFormat = "raw",
  branch = "master", onlyList = FALSE)
}
\arguments{
\item{name}{Name of the instance(s) or only parts of the name. May be an regular expression for
the instance name (excluding the file extension).}

\item{class}{Problem class. Ignored if \code{name} used.}

\item{fileFormat}{The file format of the instance. That is, the file extension of the instance.}

\item{branch}{The branch at GitHub.}

\item{onlyList}{If true don't download the instances but only list the instances (including file
path at GitHub). Useful for testing.}
}
\value{
The names of the instances.
}
\description{
Download instances from GitHub
}
\examples{
getInstance(name="SSCFLP")
getInstance(name="SSCFLP.*p6")
}
\author{
Lars Relund \email{lars@relund.dk}
}
