% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getInstanceFolders}
\alias{getInstanceFolders}
\title{Get instance folders (subfolders in the instances folder)}
\usage{
getInstanceFolders(class = NULL, branch = "master", local = FALSE)
}
\arguments{
\item{class}{Problem classes considered (if NULL then all).}

\item{branch}{The branch at GitHub.}

\item{local}{Use local repo.}
}
\value{
The folder names.
}
\description{
Get instance folders (subfolders in the instances folder)
}
\examples{
getInstanceFolders()
getInstanceFolders(local = TRUE)
getInstanceFolders(class = "TSP")
}
\author{
Lars Relund \email{lars@relund.dk}
}
