% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getInstanceInfo}
\alias{getInstanceInfo}
\title{Get info about the instances.}
\usage{
getInstanceInfo(class = NULL, folder = NULL, branch = "master",
  local = FALSE)
}
\arguments{
\item{class}{Problem class of interest (if NULL consider all classes).}

\item{folder}{Folder(s) of interest (if NULL consider all folders).}

\item{branch}{The branch at GitHub.}

\item{local}{Use local repo.}
}
\value{
A list (invisible).
}
\description{
Get info about the instances.
}
\examples{
getInstanceInfo(local = T)
getInstanceInfo(local = F)
}
\author{
Lars Relund \email{lars@relund.dk}
}
