% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getInstanceList}
\alias{getInstanceList}
\title{Get instance files (use \code{metaInstances.json})}
\usage{
getInstanceList(name = "", class = NULL, contribution = NULL, local = FALSE)
}
\arguments{
\item{name}{Name of the file(s) or only parts of the name. May be an regular expression.}

\item{class}{Problem class. Ignored if \code{name} used.}

\item{contribution}{Name of the contribution (without prefix MOrepo-). If NULL consider all
folders.}

\item{local}{Use local repo.}
}
\value{
The names of the files (including file path)
}
\description{
Get instance files (use \code{metaInstances.json})
}
\examples{
getInstanceList("SSCFLP.*p64")
getInstanceList(class = "Facility location")
getInstanceList(contribution="Tuyttens00")
}
\author{
Lars Relund \email{lars@relund.dk}
}
