% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getMaintainers}
\alias{getMaintainers}
\title{Get maintainers of stuff at MOrepo.}
\usage{
getMaintainers(branch = "master", local = FALSE)
}
\arguments{
\item{branch}{The branch at GitHub.}

\item{local}{Use local repo.}
}
\value{
A vector of maintainers (invisible).
}
\description{
Get maintainers of stuff at MOrepo.
}
\examples{
getMaintainers()
}
\author{
Lars Relund \email{lars@relund.dk}
}
