% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getMetaInstances}
\alias{getMetaInstances}
\title{Get data frame over all instances at MOrepo}
\usage{
getMetaInstances()
}
\description{
Use \code{metaInstances.json} as source.
}
\examples{
getMetaInstances()
}
\author{
Lars Relund \email{lars@relund.dk}
}
