% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getProblemClasses}
\alias{getProblemClasses}
\title{Get all the problem classes}
\usage{
getProblemClasses(local = FALSE, contribution = NULL, results = FALSE)
}
\arguments{
\item{local}{Use local repo.}

\item{contribution}{Consider a specific contribution.}

\item{results}{If TRUE return only problem classes with results.}
}
\value{
All problem classes.
}
\description{
Get all the problem classes
}
\examples{
getProblemClasses()
getProblemClasses(results = TRUE)
getProblemClasses(local = T)
getProblemClasses(contribution = "Pedersen08")
}
\author{
Lars Relund \email{lars@relund.dk}
}
