% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getRepoPath}
\alias{getRepoPath}
\title{Get urls of the sub-repositories of MOrepo.}
\usage{
getRepoPath(local = FALSE)
}
\arguments{
\item{local}{Use local repo (urls becomes paths).}
}
\value{
A vector of strings.
}
\description{
Get urls of the sub-repositories of MOrepo.
}
\examples{
getRepoPath()
}
\author{
Lars Relund \email{lars@relund.dk}
}
