% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getResultInfo}
\alias{getResultInfo}
\title{Get info about the results.}
\usage{
getResultInfo(class = NULL, contribution = NULL, local = FALSE,
  silent = FALSE, withLinks = FALSE)
}
\arguments{
\item{class}{Problem class of interest (if NULL consider all classes).}

\item{contribution}{Name of the contribution (without prefix MOrepo-). If NULL consider all folders.}

\item{local}{Use local repos. Assume that repositories are placed in the father folder of the
current working dir.}

\item{silent}{If true no output.}

\item{withLinks}{Output will be with markdown links.}
}
\value{
A list (invisible).
}
\description{
Get info about the results.
}
\examples{
getResultInfo()
getResultInfo(withLinks = TRUE)
getResultInfo(class="Assignment")
getResultInfo(contribution = c("Pedersen08", "Gadegaard16"))
}
\author{
Lars Relund \email{lars@relund.dk}
}
