% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getResultList}
\alias{getResultList}
\title{Get result files (use \code{metaResults.json})}
\usage{
getResultList(name = "", class = NULL, contribution = NULL)
}
\arguments{
\item{name}{Name of the file(s) or only parts of the name. May be an regular expression.}

\item{class}{Problem class. Ignored if \code{name} used.}

\item{contribution}{Name of the contribution (without prefix MOrepo-). If NULL consider all
folders.}
}
\value{
The names of the files (including file path)
}
\description{
Get result files (use \code{metaResults.json})
}
\examples{
getInstanceList("SSCFLP.*p64")
getInstanceList(class = "Facility location")
getInstanceList(contribution="Tuyttens00")
}
\author{
Lars Relund \email{lars@relund.dk}
}
