% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_doc.R
\docType{package}
\name{MOrepoTools-package}
\alias{MOrepoTools}
\alias{MOrepoTools-package}
\title{MOrepoTools: Tools for the MOrepo at GitHub}
\description{
MOrepoTools is a R package with tools for the MOrepo at GitHub. It is not nesseary for downloading/uploading e.g. instances from MOrepo but it will make life mush easier for you.
}
\examples{
## Info about MOrepo
## Working dir must be the MOrepo dir

# Problem classes
getProblemClasses()  # problem classes with instances
getProblemClasses(results = TRUE)   # problem classes with results
getProblemClasses(contribution = "Pedersen08")  # problem classes for a specific contribution

# Info about instances
getInstanceInfo() # all contributions with instances
getInstanceInfo(class = "Assignment")   # contributions with instances for problem class Assignment
getInstanceInfo(contribution = c("Tuyttens00", "Gadegaard16"))   # info about two contributions

# Info about results
getResultInfo()   # all contributions with results
getResultInfo(withLinks = TRUE)   # add links to output
getResultInfo(class="Assignment")   # contributions with results for problem class Assignment
getResultInfo(contribution = c("Pedersen08", "Gadegaard16"))   # info about two contributions


## Download
# Download two contributions
getContributionAsZip(c("Tuyttens00", "Template"))

# Download specific instances
getInstance(name="Tuyttens.*n10", onlyList = TRUE)  # only file locations returned (no downloading)


## Plotting results
plotNDSet("results/AP/Tuyttens00_AP_n10_result.json", contribution = "Pedersen08")
plotNDSet("results/AP/Tuyttens00_AP_n10_result.json", contribution = "Pedersen08", labels = 1:16)
plotNDSet("results/AP/Tuyttens00_AP_n10_result.json", contribution = "Pedersen08",
          addTriangles = TRUE)
plotNDSet("results/AP/Tuyttens00_AP_n10_result.json", contribution = "Pedersen08", addHull = TRUE)
plotNDSet("results/AP/Tuyttens00_AP_n10_result.json", contribution = "Pedersen08", addHull = TRUE,
          addTriangles = TRUE)


## Contributing to MOrepo
# Adding a result file
points <- data.frame(z1 = c(27, 30, 31, 34, 42, 43, 49, 51), z2 = c(56, 53, 36, 33, 30, 25, 23, 9),
 type = c('se', 'us', 'se', 'us', 'us', 'us', 'us', 'se'))
createResultFile(instanceName = "Tuyttens00_AP_n05", contributionName = "Pedersen08",
   objectives = 2, points = points, card = 8, suppCard = 3, extCard = 3,
   objectiveType = c("int", "int"), direction = c("min", "min"),
   comments = "Results from the paper by Pedersen et. al (2008)", optimal = TRUE
)
modifyResultFile("Tuyttens00_AP_n05_result.json", comments = "New changed comment")

## Check your contribution
# checkContribution()

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MCDMSociety/MOrepo/misc/R/MOrepoTools}
  \item Report bugs at \url{https://github.com/MCDMSociety/MOrepo/issues}
}

}
\author{
\strong{Maintainer}: Lars Relund \email{lars@relund.dk}

}
\keyword{internal}
