% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{setMetaContributions}
\alias{setMetaContributions}
\title{Create the \code{metaContributions.json} file}
\usage{
setMetaContributions(add = NULL)
}
\arguments{
\item{add}{Contribution to add (without the prefix \verb{MOrepo-}) and then update.}
}
\description{
Note use the entry \code{repos} in \code{metaContributions.json} to find out which sub-repos to
scan and then overwrite the file.
}
\examples{
\dontrun{
MOrepoTools:::setMetaContributions()
MOrepoTools:::setMetaContributions(add = "Adelgren16")
}
}
\author{
Lars Relund \email{lars@relund.dk}
}
