% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{setMetaInstances}
\alias{setMetaInstances}
\title{Create the \code{metaInstances.json} file}
\usage{
setMetaInstances()
}
\description{
Must be run each time there is instance modifications.
}
\examples{
\dontrun{
MOrepoTools:::setMetaInstances()
}
}
\author{
Lars Relund \email{lars@relund.dk}
}
