% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{setMetaResults}
\alias{setMetaResults}
\title{Create the \code{metaResults.json} file}
\usage{
setMetaResults()
}
\description{
Must be run each time there is result modifications.
}
\examples{
\dontrun{
MOrepoTools:::setMetaResults()
}
}
\author{
Lars Relund \email{lars@relund.dk}
}
