% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{vec2String}
\alias{vec2String}
\title{Convert a vector to a string with an 'and' at the second last (if needed).}
\usage{
vec2String(v)
}
\arguments{
\item{v}{A vector.}
}
\description{
Convert a vector to a string with an 'and' at the second last (if needed).
}
\examples{
vec2String("Test")
vec2String(1:2)
vec2String(1:5)
}
\author{
Lars Relund \email{lars@relund.dk}
}
