% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{yesno}
\alias{yesno}
\title{Function of asking questions.}
\usage{
yesno(...)
}
\arguments{
\item{...}{Question string.}
}
\value{
True if yes and false otherwise.
}
\description{
Function of asking questions.
}
\author{
Lars Relund \email{lars@relund.dk}
}
