% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getFileList}
\alias{getFileList}
\title{Search repos and get a list of files}
\usage{
getFileList(name = "", subdir = "", contribution = NULL, local = FALSE)
}
\arguments{
\item{name}{Name of the file(s) or only parts of the name. May be an regular expression.}

\item{subdir}{Restricts search to a specific subfolder in each repo. The path must end with a
slash (/).}

\item{contribution}{Name of the contribution (without prefix MOrepo-). If NULL consider all folders.}

\item{local}{Use local repo.}
}
\value{
The names of the files (including file path)
}
\description{
Search repos and get a list of files
}
\examples{
getFileList("SSCFLP.*p6")
getFileList(".json")
getFileList(c(".json","ReadMe"))
getFileList("ReadMe", contribution=c("Gadegaard16", "Tuyttens00"))
getFileList("ReadMe", contribution=c("Gadegaard16", "Tuyttens00"), subdir = "instances/")
getFileList(".xml", contribution="Tuyttens00")
getFileList(".xml", contribution="Tuyttens00", local = T)
getFileList(".json", local =T)
}
\author{
Lars Relund \email{lars@relund.dk}
}
