% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getInstance}
\alias{getInstance}
\title{Download instances from GitHub}
\usage{
getInstance(name = NULL, class = NULL, fileFormat = "raw", onlyList = FALSE)
}
\arguments{
\item{name}{Name of the instance(s) or only parts of the name. May be an regular expression for
the instance name (excluding the file extension).}

\item{class}{Problem class. Ignored if \code{name} used.}

\item{fileFormat}{The file format of the instance. That is, the file extension of the instance.}

\item{onlyList}{If true don't download the instances but only list the instances (including file
path at GitHub). Useful for testing.}
}
\value{
The names of the instances.
}
\description{
Download instances from GitHub
}
\examples{
## Info about MOrepo
## Working dir must be the MOrepo dir

# Problem classes
getProblemClasses()  # problem classes with instances
getProblemClasses(results = TRUE)   # problem classes with results
getProblemClasses(contribution = "Pedersen08")  # problem classes for a specific contribution

# Info about instances
getInstanceInfo() # all contributions with instances
getInstanceInfo(class = "Assignment")   # contributions with instances for problem class Assignment
getInstanceInfo(contribution = c("Tuyttens00", "Gadegaard16"))   # info about two contributions

# Info about results
getResultInfo()   # all contributions with results
getResultInfo(withLinks = TRUE)   # add links to output
getResultInfo(class="Assignment")   # contributions with results for problem class Assignment
getResultInfo(contribution = c("Pedersen08", "Gadegaard16"))   # info about two contributions


## Download
# Download two contributions
getContributionAsZip(c("Tuyttens00", "Template"))

# Download specific instances
getInstance(name="Tuyttens.*n10", onlyList = TRUE)  # only file locations returned (no downloading)


## Plotting results
plotNDSet("results/AP/Tuyttens00_AP_n10_result.json", contribution = "Pedersen08")
plotNDSet("results/AP/Tuyttens00_AP_n10_result.json", contribution = "Pedersen08", labels = 1:16)
plotNDSet("results/AP/Tuyttens00_AP_n10_result.json", contribution = "Pedersen08",
          addTriangles = TRUE)
plotNDSet("results/AP/Tuyttens00_AP_n10_result.json", contribution = "Pedersen08", addHull = TRUE)
plotNDSet("results/AP/Tuyttens00_AP_n10_result.json", contribution = "Pedersen08", addHull = TRUE,
          addTriangles = TRUE)


## Contributing to MOrepo
# Adding a result file
points <- data.frame(z1 = c(27, 30, 31, 34, 42, 43, 49, 51), z2 = c(56, 53, 36, 33, 30, 25, 23, 9),
 type = c('se', 'us', 'se', 'us', 'us', 'us', 'us', 'se'))
createResultFile(instanceName = "Tuyttens00_AP_n05", contributionName = "Pedersen08",
   objectives = 2, points = points, card = 8, suppCard = 3, extCard = 3,
   objectiveType = c("int", "int"), direction = c("min", "min"),
   comments = "Results from the paper by Pedersen et. al (2008)", optimal = TRUE
)
modifyResultFile("Tuyttens00_AP_n05_result.json", comments = "New changed comment")

# Check your contribution
\dontrun{
checkContribution()
}
}
\author{
Lars Relund \email{lars@relund.dk}
}
