% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getInstanceInfo}
\alias{getInstanceInfo}
\title{Get info about the instances.}
\usage{
getInstanceInfo(class = NULL, contribution = NULL, local = FALSE,
  silent = FALSE)
}
\arguments{
\item{class}{Problem class of interest (if NULL consider all classes).}

\item{contribution}{Name of the contribution (without prefix MOrepo-). If NULL consider all folders.}

\item{local}{Use local repos. Assume that repositories are placed in the father folder of the
current working dir.}

\item{silent}{If true no output.}
}
\value{
A list (invisible).
}
\description{
Get info about the instances.
}
\examples{
getInstanceInfo()
getInstanceInfo(class="Facility location")
}
\author{
Lars Relund \email{lars@relund.dk}
}
