% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getMaintainers}
\alias{getMaintainers}
\title{Get maintainers of all sub-repos at MOrepo (read from the meta.json file).}
\usage{
getMaintainers(branch = "master", local = FALSE)
}
\arguments{
\item{local}{Use local repo.}
}
\value{
A vector of maintainers.
}
\description{
Get maintainers of all sub-repos at MOrepo (read from the meta.json file).
}
\examples{
getMaintainers()
getMaintainers(local = T)
}
\author{
Lars Relund \email{lars@relund.dk}
}
