% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getProblemClasses}
\alias{getProblemClasses}
\title{Get all the problem classes}
\usage{
getProblemClasses(local = FALSE)
}
\arguments{
\item{local}{Use local repo.}
}
\value{
All problem classes.
}
\description{
Get all the problem classes
}
\examples{
getProblemClasses()
getProblemClasses(local = T)
}
\author{
Lars Relund \email{lars@relund.dk}
}
