% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getProblemClasses}
\alias{getProblemClasses}
\title{Get the problem classes}
\usage{
getProblemClasses(branch = "master", local = FALSE)
}
\arguments{
\item{branch}{The branch at GitHub.}

\item{local}{Use local repo.}
}
\value{
All problem classes.
}
\description{
Get the problem classes
}
\examples{
getProblemClasses()
}
\author{
Lars Relund \email{lars@relund.dk}
}
