% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{setMetaContributions}
\alias{setMetaContributions}
\title{Create the \code{metaContributions.json} file}
\usage{
setMetaContributions()
}
\description{
Note use the entry \code{repos} in \code{metaContributions.json} to find out which sub-repos to
scan and then overwrite the file.
}
\examples{
MOrepoTools:::setMetaContributions()
}
\author{
Lars Relund \email{lars@relund.dk}
}
