% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{setMetaInstances}
\alias{setMetaInstances}
\title{Create the \code{metaInstances.json} file}
\usage{
setMetaInstances()
}
\description{
Create the \code{metaInstances.json} file
}
\examples{
setMetaInstances()
}
\author{
Lars Relund \email{lars@relund.dk}
}
